# product-api/swagger.yaml
basePath: /v2
consumes:
- application/json
- application/xml
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/mnbao1975/microservices/product-api/handlers
  Product:
    description: Product defines the structure for an API product
    properties:
      description:
        type: string
        x-go-name: Description
      id:
        description: the id for this product
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
      price:
        format: float
        type: number
        x-go-name: Price
      sku:
        type: string
        x-go-name: SKU
    required:
    - id
    type: object
    x-go-package: github.com/mnbao1975/microservices/product-api/data
host: localhost
info:
  description: Documentation Product API.
  title: Product API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
  /products/{id}:
    get:
      description: Returns a product
      operationId: listOneProduct
      parameters:
      - description: The id of product given on the URI path
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "400":
          $ref: '#/responses/errorResponse'
      tags:
      - products
produces:
- application/json
- application/xml
responses:
  errorResponse:
    description: An error returns in the response
    schema:
      $ref: '#/definitions/GenericError'
  productResponse:
    description: A product returns in the response
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of products return in the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
- https
swagger: "2.0"